🧪 Add tests for USART2_init
